Use long arithmetic in cursor bounds checks
The bounds checks added in the previous commit used int
arithmetic for valueCount * valueSize and (valueCount + 1) * 4.
With rowCount up to Integer.MAX_VALUE and valueSize up to 32,
these products can overflow to negative values, bypassing the
dataLength check entirely. Use long arithmetic for the
multiplication so overflow-wrapped values cannot slip through.
After the check passes, the value is guaranteed to fit in int
since dataLength is int, so the cast back is safe.

+++ core/src/main/java/io/questdb/cutlass/qwp/protocol/QwpDecimalColumnCursor.java
@@ -230,7 +230,7 @@ public final class QwpDecimalColumnCursor implements QwpColumnCursor {
         offset += 1;
 
         int valueCount = rowCount - nullCount;
-        int valuesSize = valueCount * valueSize;
+        long valuesSize = (long) valueCount * valueSize;
         if (offset + valuesSize > dataLength) {
             throw QwpParseException.create(
                     QwpParseException.ErrorCode.INSUFFICIENT_DATA,
@@ -238,7 +238,7 @@ public final class QwpDecimalColumnCursor implements QwpColumnCursor {
             );
         }
         this.valuesAddress = dataAddress + offset;
-        offset += valuesSize;
+        offset += (int) valuesSize;
 
         resetRowPosition();
         return offset;
