+++ core/src/main/java/io/questdb/cutlass/qwp/protocol/QwpFixedWidthColumnCursor.java
@@ -270,7 +270,7 @@ public final class QwpFixedWidthColumnCursor implements QwpColumnCursor {
         }
 
         int valueCount = rowCount - nullCount;
-        int valuesSize = valueCount * valueSize;
+        long valuesSize = (long) valueCount * valueSize;
         if (offset + valuesSize > dataLength) {
             throw QwpParseException.create(
                     QwpParseException.ErrorCode.INSUFFICIENT_DATA,
@@ -278,7 +278,7 @@ public final class QwpFixedWidthColumnCursor implements QwpColumnCursor {
             );
         }
         this.valuesAddress = dataAddress + offset;
-        offset += valuesSize;
+        offset += (int) valuesSize;
 
         resetRowPosition();
         return offset;
